fix(api): a bunch of fixes to make demo validation work
Part of ENG-599
