+++ packages/api/src/external/ehr/shared/command/patient/confirm-patient-match.ts
@@ -1,0 +1,27 @@
+import { PatientDemoData } from "@metriport/core/domain/patient";
+import { MetriportError } from "@metriport/shared";
+import { getPatientByDemo } from "../../../../../command/medical/patient/get-patient";
+
+export async function confirmPatientMatch({
+  cxId,
+  patientId,
+  demographics,
+}: {
+  cxId: string;
+  patientId: string;
+  demographics: PatientDemoData;
+}): Promise<void> {
+  const patientExists = await getPatientByDemo({ cxId, demo: demographics });
+  if (!patientExists || patientExists.id !== patientId) {
+    throw new MetriportError(
+      "Patient demographics do not match the provided Metriport patient ID",
+      undefined,
+      {
+        cxId,
+        patientId,
+        demographics: JSON.stringify(demographics),
+        message: "Please make sure you are linking the correct patient",
+      }
+    );
+  }
+}
